moved GcsStorage to gcs.py
