+++ exporter/terra/dcp_staging_client.py
@@ -1,27 +1,16 @@
 from exporter.metadata import MetadataResource, DataFile, FileChecksums
 from exporter.graph.experiment_graph import LinkSet
 from exporter.schema import SchemaService
-from exporter.terra.gcs import GcsXferStorage
-from typing import Iterable, IO, Dict, Any, Union
+from exporter.terra.gcs import GcsXferStorage, GcsStorage, Streamable
+from typing import Iterable, Dict
 
-from io import StringIO, BufferedReader
+from io import StringIO
 
 from google.cloud import storage
 from google.oauth2.service_account import Credentials
-from google.api_core.exceptions import PreconditionFailed
-from smart_open import open
 import json
-import logging
 
-from time import sleep
 from dataclasses import dataclass
-
-
-Streamable = Union[BufferedReader, StringIO, IO[Any]]
-
-
-class UploadPollingException(Exception):
-    pass
 
 
 @dataclass
@@ -52,79 +41,6 @@ class FileDescriptor:
         name = f'{data_file.uuid}_{data_file.dcp_version}_{data_file.file_name}'
         return FileDescriptor(data_file.uuid, file_metadata.dcp_version, name,
                               data_file.content_type, data_file.size, data_file.checksums)
-
-
-class GcsStorage:
-    def __init__(self, gcs_client: storage.Client, bucket_name: str, storage_prefix: str):
-        self.gcs_client = gcs_client
-        self.bucket_name = bucket_name
-        self.storage_prefix = storage_prefix
-
-        self.logger = logging.getLogger(__name__)
-        self.logger.setLevel(logging.INFO)
-
-    def file_exists(self, object_key: str) -> bool:
-        dest_key = f'{self.storage_prefix}/{object_key}'
-        staging_bucket: storage.Bucket = self.gcs_client.bucket(self.bucket_name)
-        blob: storage.Blob = staging_bucket.blob(dest_key)
-        if not blob.exists():
-            return False
-        else:
-            blob.reload()
-            return blob.metadata is not None and blob.metadata.get("export_completed", False)
-
-    def write(self, object_key: str, data_stream: Streamable):
-        try:
-            dest_key = f'{self.storage_prefix}/{object_key}'
-            staging_bucket: storage.Bucket = self.gcs_client.bucket(self.bucket_name)
-            blob: storage.Blob = staging_bucket.blob(dest_key, chunk_size=1024 * 256 * 20)
-
-            if not blob.exists():
-                blob.upload_from_file(data_stream, if_generation_match=0)
-                blob.metadata = {"export_completed": True}
-                blob.patch()
-            else:
-                self.assert_file_uploaded(object_key)
-        except PreconditionFailed as e:
-            # With if_generation_match=0, this pre-condition failure indicates that another
-            # export instance has began uploading this file. We should not attempt to upload
-            # and instead poll for its completion
-            self.assert_file_uploaded(object_key)
-
-    def move_file(self, source_key: str, object_key: str):
-        dest_key = f'{self.storage_prefix}/{object_key}'
-        staging_bucket: storage.Bucket = self.gcs_client.bucket(self.bucket_name)
-        source_blob: storage.Blob = staging_bucket.blob(source_key)
-
-        new_blob = staging_bucket.rename_blob(source_blob, dest_key)
-        new_blob.metadata = {"export_completed": True}
-        new_blob.patch()
-        return
-
-    def assert_file_uploaded(self, object_key: str):
-        dest_key = f'{self.storage_prefix}/{object_key}'
-        staging_bucket: storage.Bucket = self.gcs_client.bucket(self.bucket_name)
-        blob = staging_bucket.blob(dest_key)
-
-        one_hour_in_seconds = 60 * 60
-        one_hundred_milliseconds = 0.1
-        return self._assert_file_uploaded(blob, one_hundred_milliseconds, one_hour_in_seconds)
-
-    def _assert_file_uploaded(self, blob: storage.Blob, sleep_time: float, max_sleep_time: float):
-        if sleep_time > max_sleep_time:
-            raise UploadPollingException(f'Could not verify completed upload for blob {blob.name} within maximum '
-                                         f'wait time of {str(max_sleep_time)} seconds')
-        else:
-            sleep(sleep_time)
-            blob.reload()
-
-            export_completed = blob.metadata is not None and blob.metadata.get("export_completed")
-            if export_completed:
-                return
-            else:
-                new_sleep_time = sleep_time * 2
-                self.logger.info(f'Verifying upload of blob {blob.name}. Waiting for {str(new_sleep_time)} seconds...')
-                return self._assert_file_uploaded(blob, new_sleep_time, max_sleep_time)
 
 
 class DcpStagingException(Exception):
